Refactor commands into separate classes
There should be no change in user-facing behavior.
