+++ pom.xml
@@ -178,5 +178,5 @@
 		      <version>2.5</version>
 		    </dependency>
 	</dependencies>
-	<version>3.1.2</version>
+	<version>3.1.3-SNAPSHOT</version>
 </project>
