+++ src/com/github/catageek/ByteCart/Commands/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains command executors for various commands.
+ */
+package com.github.catageek.ByteCart.Commands;
